requirements.txt
transformers==4.11.3
torchaudio
librosa
jiwer
datasets

